// src/models/EventType.ts
/**
 * Enum for event types
 */
export enum EventType {
  BATTLE = 'Battle',
  DISCOVERY = 'Discovery',
  SOCIAL = 'Social',
  TRAVEL = 'Travel',
  QUEST = 'Quest',
  EXPLORATION = 'Exploration',
  REST = 'Rest',
  CEREMONY = 'Ceremony',
  DISASTER = 'Disaster',
  FESTIVAL = 'Festival',
  POLITICAL = 'Political',
  HISTORICAL = 'Historical',
  OTHER = 'Other'
}